requirements.txt
requests
redis
dnspython==1.15.0
splinter==0.7.5
celery==3.1.25
flower==0.7.3
pygeoip==0.3.2
gevent==1.2.1
bs4==0.0.1
beautifulsoup4==4.6.0
MySQL-python==1.2.3
